backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
python-dotenv>=1.1.0
pydantic>=2.11.7
google-genai>=1.50.1
pydantic-settings>=2.6.1

# Google Drive integration
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.154.0

# Database for persistence
sqlalchemy>=2.0.36
alembic>=1.14.0

# Security - API key encryption
cryptography>=43.0.0

# Scheduler for auto-sync
apscheduler>=3.10.4

# MCP Server integration
fastmcp>=2.13.1
httpx[socks]>=0.28.1

# CLI tool
click>=8.1.8
rich>=13.9.4
pyyaml>=6.0.2

# Testing
pytest>=8.3.4
pytest-mock>=3.14.0
